Extract SQL test helper
