+++ Jinaga.Store.SQLite.Test/QueryGeneratorTests.cs
@@ -1,8 +1,5 @@
 using FluentAssertions;
 using Jinaga.Facts;
-using Jinaga.Projections;
-using Jinaga.Store.SQLite.Builder;
-using Jinaga.Store.SQLite.Generation;
 using Jinaga.Store.SQLite.Test.Models;
 using System.Collections.Immutable;
 
@@ -17,7 +14,7 @@ public class QueryGeneratorTests
             facts.OfType<Department>()
                 .Where(department => department.company == company)
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -42,7 +39,7 @@ public class QueryGeneratorTests
             facts.OfType<Company>()
                 .Where(company => company == department.company)
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -73,7 +70,7 @@ public class QueryGeneratorTests
             ).Any<ProjectDeleted>()
             select project
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -116,7 +113,7 @@ public class QueryGeneratorTests
             ).Any<ProjectDeleted>()
             select project
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -164,7 +161,7 @@ public class QueryGeneratorTests
             ).Any<ProjectDeleted>()
             select project
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -215,7 +212,7 @@ public class QueryGeneratorTests
                     project.department == department)
             }
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -264,7 +261,7 @@ public class QueryGeneratorTests
                 next.prior.Contains(name)
             select next
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -302,7 +299,7 @@ public class QueryGeneratorTests
                     next.prior.Contains(name))
             select name
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -342,7 +339,7 @@ public class QueryGeneratorTests
                 name.prior.Contains(prior)
             select prior
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -380,7 +377,7 @@ public class QueryGeneratorTests
             where assignment.user == user
             select assignment
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -425,7 +422,7 @@ public class QueryGeneratorTests
             where assignment.project == project
             select assignment
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -470,7 +467,7 @@ public class QueryGeneratorTests
                     assignment.user == user)
             select project
         );
-        SqlQueryTree sqlQueryTree = SqlFor(specification);
+        SqlQueryTree sqlQueryTree = specification.ToSql();
 
         sqlQueryTree.SqlQuery.Sql.Should().Be(
             "SELECT " +
@@ -504,186 +501,5 @@ public class QueryGeneratorTests
             ") " +
             "ORDER BY f3.fact_id ASC"
         );
-    }
-
-    private SqlQueryTree SqlFor(Specification specification)
-    {
-        var factTypes = GetAllFactTypes(specification);
-        var roleMap = GetAllRoles(specification, factTypes);
-
-        var givenTuple = specification.Givens
-            .Select((given, index) => (
-                name: given.Label.Name,
-                reference: new FactReference(given.Label.Type, $"{index + 1001}")
-            ))
-            .Aggregate(FactReferenceTuple.Empty, (tuple, item) => tuple.Add(item.name, item.reference));
-
-        var descriptionBuilder = new ResultDescriptionBuilder(factTypes, roleMap);
-        string str = specification.ToString();
-        var description = descriptionBuilder.Build(givenTuple, specification);
-
-        var sqlQueryTree = SqlGenerator.CreateSqlQueryTree(description);
-        return sqlQueryTree;
-    }
-
-    private ImmutableDictionary<string, int> GetAllFactTypes(Specification specification)
-    {
-        return GetAllFactTypesFromSpecification(specification)
-            .Select((name, index) => KeyValuePair.Create(name, index + 1))
-            .ToImmutableDictionary();
-    }
-
-    private IEnumerable<string> GetAllFactTypesFromSpecification(Specification specification)
-    {
-        var factTypeNames = specification.Givens
-            .Select(g => g.Label.Type)
-            .ToImmutableList();
-        factTypeNames = factTypeNames.AddRange(GetAllFactTypesFromMatches(specification.Matches));
-        if (specification.Projection is CompoundProjection compoundProjection)
-        {
-            factTypeNames = factTypeNames.AddRange(GetAllFactTypesFromProjection(compoundProjection));
-        }
-        return factTypeNames.Distinct();
-    }
-
-    private IEnumerable<string> GetAllFactTypesFromMatches(ImmutableList<Match> matches)
-    {
-        var factTypeNames = ImmutableList<string>.Empty;
-        foreach (var match in matches)
-        {
-            factTypeNames = factTypeNames.Add(match.Unknown.Type);
-            foreach (var pathCondition in match.PathConditions)
-            {
-                factTypeNames = factTypeNames.AddRange(
-                    pathCondition.RolesLeft.Select(role =>
-                        role.TargetType));
-                factTypeNames = factTypeNames.AddRange(
-                    pathCondition.RolesRight.Select(role =>
-                        role.TargetType));
-            }
-            foreach (var existentialCondition in match.ExistentialConditions)
-            {
-                factTypeNames = factTypeNames.AddRange(
-                    GetAllFactTypesFromMatches(existentialCondition.Matches));
-            }
-        }
-        return factTypeNames.Distinct();
-    }
-
-    private IEnumerable<string> GetAllFactTypesFromProjection(CompoundProjection compoundProjection)
-    {
-        var factTypeNames = ImmutableList<string>.Empty;
-
-        foreach (var name in compoundProjection.Names)
-        {
-            var projection = compoundProjection.GetProjection(name);
-            if (projection is CompoundProjection nestedCompoundProjection)
-            {
-                factTypeNames = factTypeNames.AddRange(
-                    GetAllFactTypesFromProjection(nestedCompoundProjection));
-            }
-            else if (projection is CollectionProjection collectionProjection)
-            {
-                factTypeNames = factTypeNames.AddRange(
-                    GetAllFactTypesFromMatches(collectionProjection.Matches));
-                if (collectionProjection.Projection is CompoundProjection nestedCompoundProjection2)
-                {
-                    factTypeNames = factTypeNames.AddRange(
-                        GetAllFactTypesFromProjection(nestedCompoundProjection2));
-                }
-            }
-        }
-        return factTypeNames.Distinct();
-    }
-
-    private ImmutableDictionary<int, ImmutableDictionary<string, int>> GetAllRoles(Specification specification, ImmutableDictionary<string, int> factTypes)
-    {
-        var distinctRoles = GetAllRolesFromSpecification(specification)
-            .GroupBy(pair => pair.factType, pair => pair.roleName)
-            .SelectMany(group => group.Distinct().Select(roleName => new
-            {
-                FactType = group.Key,
-                RoleName = roleName
-            }));
-        var rolesByFactTypeId = distinctRoles
-            .Select((pair, index) => new
-            {
-                FactTypeId = factTypes[pair.FactType],
-                RoleName = pair.RoleName,
-                RoleId = index + 1
-            })
-            .GroupBy(role => role.FactTypeId, role => KeyValuePair.Create(
-                role.RoleName, role.RoleId
-            ))
-            .Select(pair => KeyValuePair.Create(pair.Key, pair.ToImmutableDictionary()))
-            .ToImmutableDictionary();
-        return rolesByFactTypeId;
-    }
-
-    private IEnumerable<(string factType, string roleName)> GetAllRolesFromSpecification(Specification specification)
-    {
-        var typesByLabel = specification.Givens
-            .Select(g => KeyValuePair.Create(g.Label.Name, g.Label.Type))
-            .ToImmutableDictionary();
-        typesByLabel = typesByLabel.AddRange(specification.Matches
-            .Select(match => KeyValuePair.Create(match.Unknown.Name, match.Unknown.Type)));
-        var roles = GetAllRolesFromMatches(typesByLabel, specification.Matches).ToImmutableList();
-        roles = roles.AddRange(GetAllRolesFromProjection(typesByLabel, specification.Projection));
-        return roles;
-    }
-
-    private IEnumerable<(string factType, string roleName)> GetAllRolesFromMatches(ImmutableDictionary<string, string> typesByLabel, ImmutableList<Match> matches)
-    {
-        var roles = ImmutableList<(string factType, string roleName)>.Empty;
-        foreach (var match in matches)
-        {
-            foreach (var pathCondition in match.PathConditions)
-            {
-                var type = match.Unknown.Type;
-                foreach (var role in pathCondition.RolesLeft)
-                {
-                    roles = roles.Add((type, role.Name));
-                    type = role.TargetType;
-                }
-                type = typesByLabel[pathCondition.LabelRight];
-                foreach (var role in pathCondition.RolesRight)
-                {
-                    roles = roles.Add((type, role.Name));
-                    type = role.TargetType;
-                }
-            }
-            foreach (var existentialCondition in match.ExistentialConditions)
-            {
-                typesByLabel = typesByLabel.AddRange(existentialCondition.Matches
-                    .Select(match => KeyValuePair.Create(match.Unknown.Name, match.Unknown.Type)));
-                roles = roles.AddRange(GetAllRolesFromMatches(typesByLabel, existentialCondition.Matches));
-            }
-        }
-        return roles;
-    }
-
-    private IEnumerable<(string factType, string roleName)> GetAllRolesFromProjection(ImmutableDictionary<string, string> typesByLabel, Projection projection)
-    {
-        var roles = ImmutableList<(string factType, string roleName)>.Empty;
-        if (projection is CompoundProjection compoundProjection)
-        {
-            foreach (var name in compoundProjection.Names)
-            {
-                var nestedProjection = compoundProjection.GetProjection(name);
-                if (nestedProjection is CompoundProjection nestedCompoundProjection)
-                {
-                    roles = roles.AddRange(GetAllRolesFromProjection(typesByLabel, nestedCompoundProjection));
-                }
-                else if (nestedProjection is CollectionProjection collectionProjection)
-                {
-                    roles = roles.AddRange(GetAllRolesFromMatches(typesByLabel, collectionProjection.Matches));
-                    if (collectionProjection.Projection is CompoundProjection nestedCompoundProjection2)
-                    {
-                        roles = roles.AddRange(GetAllRolesFromProjection(typesByLabel, nestedCompoundProjection2));
-                    }
-                }
-            }
-        }
-        return roles;
     }
 }
